convert to Typescript
Please ignore the fact that this is all in one commit

When the project was first made, I used a template from the Phaser devs. They happened to have an equivalent template for TypeScript. So the hardest part of the work was likely already done for me.

What I ended up doing was going through the files, Converting to .ts and fixing problems that arose. A lot of this was type declaration issues, but some other problems came up, like import troubles.

For anyone who is using this version for the first time now, you WILL need to do npm install